pyproject.toml
[tool.poetry]
name = "fresh-wallets-parser"
version = "0.1.0"
description = ""
authors = ["Андрей Кагадий <kagadi.a@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
telethon = "^1.33.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
